requirements.txt
# core dependencies
discord.py==2.3.2
google-generativeai==0.3.2
python-dotenv==1.0.0

# web server
aiohttp==3.9.1

# logging
colorlog==6.7.0

# testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

aiohappyeyeballs==2.4.8
aiosignal==1.3.2
attrs==25.1.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
yarl==1.18.3
typing-extensions==4.9.0
protobuf==4.25.3
grpcio==1.62.1
grpcio-status==1.62.1
